Initial import of the pkg_comp-2.0 sources
pkg_comp automates the build of pkgsrc binary packages in a sandbox.

pkg_comp is a tool to automate the maintenance of a set of binary
packages, built from the pkgsrc packaging system, by means of a single
command and configuration file.

pkg_comp orchestrates VCS tools such as CVS to fetch the initial pkgsrc
tree and to keep it up to date; the creation of a fresh sandbox to build
packages in; the pkgsrc bootstrapping process; and the use of the pbulk
infrastructure to build the desired set of packages in an optimal
manner.

As you may notice from the copyright notices, this code has been in the
works since 2013 (and pkg_comp 1.x itself has existed since 2002!).  The
code for this rewrite had not been clean enough for publication until
now (not even its git history), hence the clean reimport in this commit.

This is not an official Google product.