Separate the pbulk and pkg binary packages
The PACKAGES variable, which is configurable by the user, is intended to
match the variable of the same name exposed by pkgsrc.  As such, we
should respect the layout of the resulting tree (i.e. there ought to be
an All subdirectory including the actual binary packages, without any
intermediate components).

Therefore, ensure PACKAGES only refers to the location where binary
packages for the target environment will be placed, and add a new
PBULK_PACKAGES variable that points to the repository for the internal
pbulk-related packages.