Add the pkg_comp_git module
This module provides convenience functions to fetch sources from git
and is the counterpart to shtk_cvs(3).  The two offer a very similar
interface, and pkg_comp_git should eventually be moved to shtk once
it proves stable.  (git is quite complex and it may require a few
iterations to get the automation here right for the more general use
cases.)