Make the auto command reuse existing sandboxes
I keep hitting cases where I run the auto command and the build fails
soon after because there was a sandbox left behind due to some prior
manual action.  This is turning to be quite annoying, so make auto
reuse the sandbox instead.

Reusing the sandbox may be bad for long-term stability as the sandbox
could become stale.  Will revisit this if it turns out to be a bigger
problem.