Fix expand_packages for repeated subdirectories
It is possible for a directory with a package's name to appear somewhere
else in the pkgsrc tree outside of the package's category.  For example,
the filesystem/fuse-unionfs package downloads its distfiles under
distfiles/fuse-unionfs, and that would confuse our expand_packages if
the distfiles were stored within pkgsrc (as is the default).

Make sure the candidates we consider for expansion contain a Makefile
within them, which is a better heuristic to find packages.