Allow bootstrapping if pkg-config is not present
Download and install pkgconf if we cannot find neither pkgconf nor
pkg-config during the bootstrap process.  These tools are not available
in clean NetBSD and macOS systems so doing this simplifies things.

Note that pkgconf is installed into a temporary directory separate from
the given prefix because there is no need to leave this build-only tool
installed on the user's machine.

Fixes #5.