Explicitly configure Git for tests
Tests run with a custom HOME directory so no .gitconfig is available.
This causes tests to fail sometimes, so explicitly initialize the
minimum contents to let Git work.

In theory, Git can autoconfigure user.email and user.name... but this
fails in Travis because the machines have no domain name.