Make integration tests to run on Travis
To do this, change the integration tests so that pkgsrcdir has to point
to a local Git clone of pkgsrc.  This local clone is used "as is" for
all existing tests, but new tests to verify the "fetch" phase of pkg_comp
use this local clone as the remote repository.

This makes the Travis builds only have to fetch pkgsrc from a remote site
once, which is costly.