Expose the pbulk logs outside of the sandbox
Add a new PBULK_LOG configuration variable that points to a directory
outside of the sandbox and that is used to record all pbulk build logs.
These logs are necessary to troubleshoot any build failures that happen
during the build, and are also useful under successful conditions to
get summaries about what was built.

Fixes #3.