Allow using unreleased versions in packaging tools
Extend bootstrap.sh so that we can pass unreleased versions of certain
packages to the script.  Then modify make-macos-pkg.sh so that it can
use this feature from bootstrap.sh.

These changes are to support testing upcoming versions of the pkg_comp
with sandboxctl combination without actually having to publish distfiles
that may still contain problems.